+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ CHANGELOG
 
 * Added: Permissions Helper added for more detailed permissions support.
 * Changed: ScanSettings model updated with acl.
+* Added: Model TargetGroup.acls
+* Added: Model ScanSetting properties: starttime, rrules, timezone, file_targets, and launch_now
 
 1.0.0
 ==========